Implemented test for well-formedness. Executing on test network gives no errors. (But absence of errors is not guaranteed)
Made interface for nodes in a tree that can be visited (TreeHost<HostType>) as well as an interface for visiting nodes in a tree (TreeVisitor<ReturnType, TreeType>). Both are generic, and use generic types.

Changed Behaviour.java from interface to abstract class that has a concrete implementation of TreeHost<Behaviour>.

Made and implemented WellFormedness.java based heavily on the Kotlin implementation, although restructured and adapted to Java typing. It has a static method that internally creates instances of its own private class that implements TreeVisitor<Boolean, Behaviour>.

Changed Extraction.java to test for well-formedness, returning null early if the test fails.