Added Benchmarks.kt It was decided that for now that Benchmarks remains written in Kotlin, as rewriting it to Java is difficult and time consuming.
Added benchmark option in ExtractionTesting. It crashes for some tests because it runs out of memory or stack, but it is likely because my machine do not have enough.

Added Bisimulation.java

Made class Interaction have a setContinuation() function, as it was needed in Bisimulation. The part where it is needed might be implemented wrong.

Made all classes in Label.java public.