Made almost all AST node classes in extraction.choreography and extraction.network have final fields. The exceptions are the »main« field of ProcessTerm, and »processes« of Network.
Made the copy() function of almost all AST nodes in extraction.network return »this« instead of an actual copy. The exceptions are ProcessTerm and Network which needs actual copies to function properly. Might be doable to those as well by making the non-final fields in ProcessTerm and Network final, and make new copies whenever their fields need modification, but I see no gains for doing this.